requirements.txt
# Minimal working dependencies
aiohttp==3.8.6
feedparser==6.0.10
textblob==0.17.1
numpy==1.24.3
pyyaml==6.0.1
python-dotenv==1.0.0

# Alpaca Trading (will install compatible pydantic)
alpaca-py==0.9.0

# Basic utilities
requests==2.31.0
python-dateutil==2.8.2
psutil==5.9.5

# Observability and Monitoring
prometheus_client==0.17.1

# Machine Learning for Sentiment Analysis
scikit-learn==1.3.0
joblib==1.3.2